Fix contact and chat and donate issues

+++ app/src/main/java/ir/kindnesswall/data/model/ChatModel.kt
@@ -1,3 +1,3 @@
 package ir.kindnesswall.data.model
 
-    data class ChatModel(var contactId: Int = 0, var userId: Int = 0, var chatId: Int = 0)
+data class ChatModel(var contactId: Long = 0, var userId: Long = 0, var chatId: Long = 0)
